+++ src/is_prime_internal.h
@@ -7,12 +7,12 @@ using namespace Rcpp;
 
 
 // Helper function to check if a number is prime
-bool is_prime_internal(long long num) {
+bool is_prime_internal(int num) {
   if (num <= 1) return false;
   if (num == 2 || num == 3) return true;
   if (num % 2 == 0 || num % 3 == 0) return false;
 
-  for (long long i = 5; i * i <= num; i += 6) {
+  for (int i = 5; i * i <= num; i += 6) {
     if (num % i == 0 || num % (i + 2) == 0) return false;
   }
